Web service crud was overwriting Java S classes even if they already exist.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@8693 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ xCrudGeneratorWebService/Form1.cs
@@ -106,9 +106,11 @@ namespace xCrudGeneratorWebService {
 				WriteAllSerializing(strb,className,fields);
 				File.WriteAllText(Path.Combine(SerialDir,className+".cs"),strb.ToString());
 				strb.Clear();
-				WriteAlljavaDataInterface(strb,className);
-				File.WriteAllText(Path.Combine(JavaSClassesDir,GetSname(className+docForjava)+".java"),strb.ToString());
-				strb.Clear();
+				if(!File.Exists(Path.Combine(JavaSClassesDir,GetSname(className+docForjava)+".java"))) {//Only create a java S class if one does not already exist.
+					WriteAlljavaDataInterface(strb,className);
+					File.WriteAllText(Path.Combine(JavaSClassesDir,GetSname(className+docForjava)+".java"),strb.ToString());
+					strb.Clear();
+				}
 				WriteAlljavaTableTypes(strb,className,fields);
 				File.WriteAllText(Path.Combine(JavaTableTypesDir,className+docForjava+".java"),strb.ToString());
 				#region DtoMethods OpenDentalClasses for Serializing and Deserializing
